feat: ux/ui changes pie chart series config
